Move OsuTabDropdown to it's own file
