+++ osu.Game/Graphics/UserInterface/OsuTabControl.cs
@@ -39,7 +39,7 @@ namespace osu.Game.Graphics.UserInterface
 
         private readonly Box strip;
 
-        protected override Dropdown<T> CreateDropdown() => new OsuTabDropdown();
+        protected override Dropdown<T> CreateDropdown() => new OsuTabDropdown<T>();
 
         protected override TabItem<T> CreateTabItem(T value) => new OsuTabItem(value);
 
@@ -179,101 +179,6 @@ namespace osu.Game.Graphics.UserInterface
             protected override void OnActivated() => fadeActive();
 
             protected override void OnDeactivated() => fadeInactive();
-        }
-
-        // todo: this needs to go
-        private class OsuTabDropdown : OsuDropdown<T>
-        {
-            public OsuTabDropdown()
-            {
-                RelativeSizeAxes = Axes.X;
-            }
-
-            protected override DropdownMenu CreateMenu() => new OsuTabDropdownMenu();
-
-            protected override DropdownHeader CreateHeader() => new OsuTabDropdownHeader
-            {
-                Anchor = Anchor.TopRight,
-                Origin = Anchor.TopRight
-            };
-
-            private class OsuTabDropdownMenu : OsuDropdownMenu
-            {
-                public OsuTabDropdownMenu()
-                {
-                    Anchor = Anchor.TopRight;
-                    Origin = Anchor.TopRight;
-
-                    BackgroundColour = Color4.Black.Opacity(0.7f);
-                    MaxHeight = 400;
-                }
-
-                protected override DrawableDropdownMenuItem CreateDrawableDropdownMenuItem(MenuItem item) => new DrawableOsuTabDropdownMenuItem(item) { AccentColour = AccentColour };
-
-                private class DrawableOsuTabDropdownMenuItem : DrawableOsuDropdownMenuItem
-                {
-                    public DrawableOsuTabDropdownMenuItem(MenuItem item)
-                        : base(item)
-                    {
-                        ForegroundColourHover = Color4.Black;
-                    }
-                }
-            }
-
-            protected class OsuTabDropdownHeader : OsuDropdownHeader
-            {
-                public override Color4 AccentColour
-                {
-                    get => base.AccentColour;
-                    set
-                    {
-                        base.AccentColour = value;
-                        Foreground.Colour = value;
-                    }
-                }
-
-                public OsuTabDropdownHeader()
-                {
-                    RelativeSizeAxes = Axes.None;
-                    AutoSizeAxes = Axes.X;
-
-                    BackgroundColour = Color4.Black.Opacity(0.5f);
-
-                    Background.Height = 0.5f;
-                    Background.CornerRadius = 5;
-                    Background.Masking = true;
-
-                    Foreground.RelativeSizeAxes = Axes.None;
-                    Foreground.AutoSizeAxes = Axes.X;
-                    Foreground.RelativeSizeAxes = Axes.Y;
-                    Foreground.Margin = new MarginPadding(5);
-
-                    Foreground.Children = new Drawable[]
-                    {
-                        new SpriteIcon
-                        {
-                            Icon = FontAwesome.Solid.EllipsisH,
-                            Size = new Vector2(14),
-                            Origin = Anchor.Centre,
-                            Anchor = Anchor.Centre,
-                        }
-                    };
-
-                    Padding = new MarginPadding { Left = 5, Right = 5 };
-                }
-
-                protected override bool OnHover(HoverEvent e)
-                {
-                    Foreground.Colour = BackgroundColour;
-                    return base.OnHover(e);
-                }
-
-                protected override void OnHoverLost(HoverLostEvent e)
-                {
-                    Foreground.Colour = BackgroundColourHover;
-                    base.OnHoverLost(e);
-                }
-            }
         }
     }
 }
